修改商品单价数据类型

+++ toucan-shopping-modules/order/toucan-shopping-order-entity-vo/src/main/java/com/toucan/shopping/modules/order/entity/OrderItem.java
@@ -5,6 +5,7 @@ import com.alibaba.fastjson.serializer.ToStringSerializer;
 import com.fasterxml.jackson.annotation.JsonFormat;
 import lombok.Data;
 
+import java.math.BigDecimal;
 import java.util.Date;
 
 /**
@@ -28,9 +29,9 @@ public class OrderItem {
     private Integer sellerStatus; //卖家备货状态 0备货中 1备货完成 2缺货
     private Integer buyerStatus; //买家状态 0待收货 1已收货 2换货 3退货
     private Integer productNum; //购买商品数量
-    private Double productPrice; //商品单价
-    private Double orderItemAmount; //订单单项总金额
-    private Double deliveryMoney; //配送费用
+    private BigDecimal productPrice; //商品单价
+    private BigDecimal orderItemAmount; //订单单项总金额
+    private BigDecimal deliveryMoney; //配送费用
     private Date deliveryReceiveTime; //收货时间
     private Date deliveryFinishTime; //配送人员完成时间
     private Date sellerFinishTime; //卖家完成时间
